+++ bin/node/executor/Cargo.toml
@@ -6,32 +6,34 @@ description = "Substrate node implementation in Rust."
 edition = "2018"
 
 [dependencies]
-trie-root = "0.15.2"
 codec = { package = "parity-scale-codec", version = "1.0.0" }
-sp-io = { version = "2.0.0", path = "../../../primitives/io" }
-sp-state-machine = { version = "2.0.0", path = "../../../primitives/state-machine" }
-sc-executor = { version = "2.0.0", path = "../../../client/executor" }
-sp-core = { version = "2.0.0", path = "../../../primitives/core" }
-sp-trie = { version = "2.0.0", path = "../../../primitives/trie" }
 node-primitives = { version = "2.0.0", path = "../primitives" }
 node-runtime = { version = "2.0.0", path = "../runtime" }
+sc-executor = { version = "2.0.0", path = "../../../client/executor" }
+sp-core = { version = "2.0.0", path = "../../../primitives/core" }
+sp-io = { version = "2.0.0", path = "../../../primitives/io" }
+sp-state-machine = { version = "2.0.0", path = "../../../primitives/state-machine" }
+sp-trie = { version = "2.0.0", path = "../../../primitives/trie" }
+trie-root = "0.15.2"
 
 [dev-dependencies]
-node-testing = { version = "2.0.0", path = "../testing" }
-substrate-test-client = { version = "2.0.0", path = "../../../test-utils/client" }
-sp-runtime = { version = "2.0.0", path = "../../../primitives/runtime" }
+criterion = "0.3.0"
 frame-support = { version = "2.0.0", path = "../../../frame/support" }
-pallet-balances = { version = "2.0.0", path = "../../../frame/balances" }
-pallet-transaction-payment = { version = "2.0.0", path = "../../../frame/transaction-payment" }
-pallet-session = { version = "2.0.0", path = "../../../frame/session" }
 frame-system = { version = "2.0.0", path = "../../../frame/system" }
-pallet-timestamp = { version = "2.0.0", path = "../../../frame/timestamp" }
-pallet-treasury = { version = "2.0.0", path = "../../../frame/treasury" }
+node-testing = { version = "2.0.0", path = "../testing" }
+pallet-balances = { version = "2.0.0", path = "../../../frame/balances" }
 pallet-contracts = { version = "2.0.0", path = "../../../frame/contracts" }
 pallet-grandpa = { version = "2.0.0", path = "../../../frame/grandpa" }
+pallet-im-online = { version = "2.0.0", path = "../../../frame/im-online" }
 pallet-indices = { version = "2.0.0", path = "../../../frame/indices" }
+pallet-session = { version = "2.0.0", path = "../../../frame/session" }
+pallet-timestamp = { version = "2.0.0", path = "../../../frame/timestamp" }
+pallet-transaction-payment = { version = "2.0.0", path = "../../../frame/transaction-payment" }
+pallet-treasury = { version = "2.0.0", path = "../../../frame/treasury" }
+sp-application-crypto = { version = "2.0.0", path = "../../../primitives/application-crypto" }
+sp-runtime = { version = "2.0.0", path = "../../../primitives/runtime" }
+substrate-test-client = { version = "2.0.0", path = "../../../test-utils/client" }
 wabt = "0.9.2"
-criterion = "0.3.0"
 
 [features]
 wasmtime = [
