+++ primitives/offchain/src/lib.rs
@@ -26,9 +26,15 @@ pub const STORAGE_PREFIX: &[u8] = b"storage";
 
 sp_api::decl_runtime_apis! {
 	/// The offchain worker api.
+	#[api_version(2)]
 	pub trait OffchainWorkerApi {
 		/// Starts the off-chain task for given block number.
 		#[skip_initialize_block]
+		#[changed_in(2)]
 		fn offchain_worker(number: NumberFor<Block>);
+
+		/// Starts the off-chain task for given block header.
+		#[skip_initialize_block]
+		fn offchain_worker(header: &Block::Header);
 	}
 }
